+++ docs/changelog.md
@@ -26,6 +26,7 @@ The way it's handling is so distinct that keeping the two implementation does no
 - Add pid in starting log message.
 - Add NO_COLOR env var support
 - Add -C option to specify the working folder.
+- Accept DSN in CLI argument.